Refactored into own module
This class uses openCV's face detection and tries to keep track of faces over time. At the moment features are fairly limited, but eventually it should allow for actions to be triggered when faces are temporarily hidden, or when a specific motion is captured. Still needs some refactoring to remove duplicate loops etc.